remove code from services to config
